+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1
 - '/addonlocale <locale>' is now accepted as an alias for '/addonlocale set <locale>'.
+- The list of locales printed by `/addonlocale set' is now sorted.
 - Updated TOC for patch 10.0.2.
 
 ## 1.0